move functions to box_util add test
