+++ mmocr/utils/__init__.py
@@ -1,5 +1,6 @@
 from mmcv.utils import Registry, build_from_cfg
 
+from .box_util import is_on_same_line, stitch_boxes_into_lines
 from .check_argument import (equal_len, is_2dlist, is_3dlist, is_ndarray_list,
                              is_none_or_type, is_type_list, valid_boundary)
 from .collect_env import collect_env
@@ -14,5 +15,5 @@ __all__ = [
     'is_3dlist', 'is_ndarray_list', 'is_type_list', 'is_none_or_type',
     'equal_len', 'is_2dlist', 'valid_boundary', 'lmdb_converter',
     'drop_orientation', 'convert_annotations', 'is_not_png', 'list_to_file',
-    'list_from_file'
+    'list_from_file', 'is_on_same_line', 'stitch_boxes_into_lines'
 ]
